requirements.txt
pyyaml==6.0.0
langchain==0.0.329
bs4==0.0.2
lxml==4.9.3
tqdm==4.64.1
regex==2022.7.9
opensearch-py==2.4.2
boto3>=1.26.73
pandas==1.5.3
Pillow==10.2.0  
streamlit==1.30.0
# torch==1.13.1
# torchvision==0.14.1
requests-aws4auth==1.2.3
botocore==1.34.37
